+++ test/cli/test_prune.py
@@ -6,10 +6,10 @@ import xarray as xr
 
 from test.cli.helpers import CliTest
 from xcube.cli.prune import _delete_block_file
-from xcube.core.new import new_cube
-from xcube.core.readwrite import write_cube
-from xcube.core.verify import assert_cube
 from xcube.core.dsio import rimraf
+from xcube.core.dsio import write_cube
+from xcube.core.new import new_cube
+from xcube.core.verify import assert_cube
 
 
 class PruneTest(CliTest):
