+++ test/cli/test_verify.py
@@ -2,9 +2,9 @@ import numpy as np
 import xarray as xr
 
 from test.cli.helpers import CliTest
-from xcube.core.new import new_cube
-from xcube.core.readwrite import write_cube
 from xcube.core.dsio import rimraf
+from xcube.core.dsio import write_cube
+from xcube.core.new import new_cube
 
 
 class VerifyTest(CliTest):
