+++ test/core/test_dsio.py
@@ -1,3 +1,4 @@
+import os
 import unittest
 from typing import Set
 
@@ -8,6 +9,41 @@ import xarray as xr
 from test.sampledata import new_test_dataset
 from xcube.core.dsio import DatasetIO, MemDatasetIO, Netcdf4DatasetIO, ZarrDatasetIO, CsvDatasetIO, \
     find_dataset_io, query_dataset_io
+from xcube.core.dsio import open_dataset, write_dataset
+from xcube.core.new import new_cube
+
+TEST_NC_FILE = "test.nc"
+TEST_NC_FILE_2 = "test-2.nc"
+
+
+class OpenWriteDatasetTest(unittest.TestCase):
+    def setUp(self):
+        super().setUp()
+        self.dataset = new_cube(variables=dict(precipitation=0.2, temperature=279.1))
+        self.dataset.to_netcdf(TEST_NC_FILE, mode="w")
+        self.dataset.close()
+
+    def tearDown(self):
+        self.dataset = None
+        os.remove(TEST_NC_FILE)
+        super().tearDown()
+
+    def test_open_dataset(self):
+        with open_dataset(TEST_NC_FILE) as ds:
+            self.assertIsNotNone(ds)
+            np.testing.assert_array_equal(ds.time.values, self.dataset.time.values)
+            np.testing.assert_array_equal(ds.lat.values, self.dataset.lat.values)
+            np.testing.assert_array_equal(ds.lon.values, self.dataset.lon.values)
+
+    def test_write_dataset(self):
+
+        dataset = new_cube()
+        try:
+            write_dataset(dataset, TEST_NC_FILE_2)
+            self.assertTrue(os.path.isfile(TEST_NC_FILE_2))
+        finally:
+            if os.path.isfile(TEST_NC_FILE_2):
+                os.remove(TEST_NC_FILE_2)
 
 
 # noinspection PyAbstractClass
