+++ test/webapi/test_s3util.py
@@ -4,9 +4,9 @@ import unittest
 from abc import abstractmethod, ABCMeta
 from typing import Dict
 
-from xcube.core.new import new_cube
-from xcube.core.readwrite import write_cube
 from xcube.core.dsio import rimraf
+from xcube.core.dsio import write_cube
+from xcube.core.new import new_cube
 from xcube.webapi.s3util import list_bucket_result_to_xml, list_s3_bucket_keys, list_s3_bucket_v2, list_s3_bucket_v1
 
 S3_BUCKET = os.path.join(os.path.dirname(__file__), "s3-bucket")
