+++ xcube/cli/prune.py
@@ -45,9 +45,9 @@ def prune(cube, dry_run):
 def _prune(input_path: str = None,
            dry_run: bool = False,
            monitor=None):
-    from xcube.core.readwrite import open_cube
     from xcube.core.chunk import get_empty_dataset_chunks
     from xcube.core.dsio import guess_dataset_format
+    from xcube.core.dsio import open_cube
 
     input_format = guess_dataset_format(input_path)
     if input_format != FORMAT_NAME_ZARR:
