+++ xcube/cli/resample.py
@@ -172,10 +172,10 @@ def _resample_in_time(input_path: str = None,
                       tolerance: str = None,
                       dry_run: bool = False,
                       monitor=None):
-    from xcube.core.readwrite import open_cube
-    from xcube.core.readwrite import write_cube
-    from xcube.core.resample import resample_in_time
     from xcube.core.dsio import guess_dataset_format
+    from xcube.core.dsio import open_cube
+    from xcube.core.dsio import write_cube
+    from xcube.core.resample import resample_in_time
 
     if not output_format:
         output_format = guess_dataset_format(output_path)
