+++ xcube/cli/vars2dim.py
@@ -43,7 +43,7 @@ def vars2dim(cube, variable, dim_name, output=None, format=None):
     """
 
     from xcube.core.dsio import guess_dataset_format
-    from xcube.core.readwrite import open_dataset, write_dataset
+    from xcube.core.dsio import open_dataset, write_dataset
     from xcube.core.vars2dim import vars_to_dim
     import os
